Fix search issue with author and dedup_titles both selected
There's an odd issue with PyMISP's search_index() function where it either truncates the search term at | or treats | as an OR statement. 

Either way, if you have a large DB filled with stuff like "Alienvault | Pulse Name" you'll get back results for every event that has Alienvault in the title. That is, unless you changing the | to \| right before searching, which is what this PR does.